backend: Reverted splitting into separate files.
