+++ getyour/app/backend.py
@@ -18,18 +18,23 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
 import json
 import datetime
+import requests
+import pendulum
 from enum import Enum
 from itertools import chain
-from decimal import Decimal
 import logging
-import pendulum
 import httpagentparser
 import magic
 
 from twilio.rest import Client
 from sendgrid.helpers.mail import Mail
 from sendgrid import SendGridAPIClient
+from usps import USPSApi, Address
 
+from python_http_client.exceptions import HTTPError as SendGridHTTPError
+from twilio.base.exceptions import TwilioRestException
+
+from django import http
 from django.shortcuts import reverse
 from django.contrib.auth.backends import UserModel
 from django.contrib.auth import login as django_auth_login
@@ -44,13 +49,15 @@ from app.models import (
     IQProgramRD,
     IQProgram,
     User,
+    Household,
+    AddressRD,
 )
-from app.constants import supported_content_types
+from app.constants import (
+    supported_content_types,
+    enable_calendar_year_renewal,
+    application_pages,
+)
 from logger.wrappers import LoggerWrapper
-
-from python_http_client.exceptions import HTTPError as SendGridHTTPError
-from twilio.base.exceptions import TwilioRestException
-from app.constants import enable_calendar_year_renewal, application_pages
 
 
 # Initialize logger
@@ -58,6 +65,37 @@ log = LoggerWrapper(logging.getLogger(__name__))
 
 
 form_page_number = 6
+
+
+# Use the following tag mapping for USPS standards for all functions
+tag_mapping = {
+    'Recipient': 'recipient',
+    'AddressNumber': 'address_2',
+    'AddressNumberPrefix': 'address_2',
+    'AddressNumberSuffix': 'address_2',
+    'StreetName': 'address_2',
+    'StreetNamePreDirectional': 'address_2',
+    'StreetNamePreModifier': 'address_2',
+    'StreetNamePreType': 'address_2',
+    'StreetNamePostDirectional': 'address_2',
+    'StreetNamePostModifier': 'address_2',
+    'StreetNamePostType': 'address_2',
+    'CornerOf': 'address_2',
+    'IntersectionSeparator': 'address_2',
+    'LandmarkName': 'address_2',
+    'USPSBoxGroupID': 'address_2',
+    'USPSBoxGroupType': 'address_2',
+    'USPSBoxID': 'address_2',
+    'USPSBoxType': 'address_2',
+    'BuildingName': 'address_1',
+    'OccupancyType': 'address_1',
+    'OccupancyIdentifier': 'address_1',
+    'SubaddressIdentifier': 'address_1',
+    'SubaddressType': 'address_1',
+    'PlaceName': 'city',
+    'StateName': 'state',
+    'ZipCode': 'zipcode',
+}
 
 
 class QualificationStatus(Enum):
@@ -79,6 +117,321 @@ def broadcast_sms(phone_Number):
         )
     except TwilioRestException as e:
         log.exception(e, function='broadcast_sms')
+
+
+def address_check(address_dict):
+    """
+    Check for address GMA and Connexion statuses.
+
+    Parameters
+    ----------
+    instance : dict
+        Post-USPS-validation dictionary. Usable data for this script are in
+        ['AddressValidateResponse']['Address'][...].
+
+    Returns
+    -------
+    bool
+        Whether the address is in the GMA (True, False).
+    bool
+        The status of Connexion service (True, False, None).
+
+    """
+
+    try:
+        # Gather the coordinate string for future queries
+        # Parse the 'instance' data for proper 'address_parts'
+        address_parts = "{}, {}".format(
+            address_dict['AddressValidateResponse']['Address']['Address2'],
+            address_dict['AddressValidateResponse']['Address']['Zip5'],
+        )
+        coord_string = address_lookup(address_parts)
+
+    except NameError:
+        # NameError specifies that the address is not found
+        # in City lookups and is therefore *probably* not in the IQ
+        # service area
+
+        # Log a potential error if the city is 'Fort Collins'
+        if address_dict['AddressValidateResponse']['Address']['City'].lower() == 'fort collins':    
+            log.error(
+                "Potential issue: Fort Collins address marked 'not in GMA': {}".format(
+                    address_dict['AddressValidateResponse']['Address'],
+                ),
+                function='address_check',
+            )
+
+        return (False, False)
+
+    else:
+        has_connexion = connexion_lookup(coord_string)
+        msg = 'Connexion not available or API not found' if has_connexion is None \
+            else 'Connexion available' if has_connexion \
+            else 'Connexion coming soon'
+        log.info(msg, function='address_check')
+
+        is_in_gma = gma_lookup(coord_string)
+        msg = 'Address is in GMA' if is_in_gma else 'Address is outside of GMA'
+        log.info(msg, function='address_check')
+
+        return (is_in_gma, has_connexion)
+
+
+def address_lookup(address_parts):
+    """
+    Look up the coordinates for an address to input into future queries.
+
+    Parameters
+    ----------
+    address_parts : str
+        The address parts to use for the lookup (specifically in the format
+        <address_2>, <zip code>) (e.g. "300 LAPORTE AVE, 80521", sans quotes).
+
+    Raises
+    ------
+    requests.exceptions.HTTPError
+        An issue with the lookup endpoint.
+    NameError
+        Address not found in City lookups - address is not in IQ service area.
+
+    Returns
+    -------
+    str
+        Formatted string of x,y coordinates for the address, to input in
+        future queries.
+
+    """
+
+    url = 'https://gisweb.fcgov.com/arcgis/rest/services/Geocode/Fort_Collins_Area_Address_Point_Geocoding_Service/GeocodeServer/findAddressCandidates'
+
+    payload = {
+        'f': 'pjson',
+        'Street': address_parts,
+    }
+
+    # Gather response
+    response = requests.get(url, params=payload)
+    if response.status_code != requests.codes.ok:
+        log.error(
+            f"API error {response.status_code}: {response.reason}; {response.content}",
+            function='address_lookup',
+        )
+        raise requests.exceptions.HTTPError(response.reason, response.content)
+
+    # Parse response
+    outVal = response.json()
+
+    # Since the gisweb endpoint seems to always return an HTTP 200, also check
+    # the JSON for an 'error' key
+    if 'error' in outVal:
+        errDict = outVal['error']
+        log.error(
+            f"API error {errDict['code']}: {errDict['message']}",
+            function='address_lookup',
+        )
+        raise requests.exceptions.HTTPError(errDict['code'], errDict['message'])
+
+    # Ensure candidate(s) exist and they have a decent match score
+    # Because this is how the Sales Tax lookup is architected, it should be
+    # safe to assume these are returned sorted, with best candidate first
+    if len(outVal['candidates']) > 0 and outVal['candidates'][0]['score'] > 85:
+        # Define the coordinate string to be used in future queries
+        coord_string = '{x},{y}'.format(
+            x=outVal['candidates'][0]['location']['x'],
+            y=outVal['candidates'][0]['location']['y'],
+        )
+
+    else:
+        raise NameError("Matching address not found")
+
+    return coord_string
+
+
+def connexion_lookup(coord_string):
+    """
+    Look up the Connexion service status given the coordinate string.
+
+    Parameters
+    ----------
+    coord_string : str
+        Formatted <x>,<y> string of coordinates from address_lookup().
+
+    Raises
+    ------
+    requests.exceptions.HTTPError
+        An issue with the lookup endpoint.
+    IndexError
+        Address not found in Connexion lookups - Connexion is likely to be
+        unavailable at this address.
+
+    Returns
+    -------
+    bool
+        Boolean 'status', designating True for 'service available' or False
+        for 'service will be available, but not yet' OR None for 'unavailable'
+        (probably)
+
+    TODO: Switch this to an enum if we want to keep this structure
+
+    """
+
+    url = 'https://gisweb.fcgov.com/arcgis/rest/services/FDH_Boundaries_ForPublic/MapServer/0/query'
+
+    payload = {
+        'f': 'pjson',
+        'geometryType': 'esriGeometryPoint',
+        'geometry': coord_string,
+    }
+
+    try:
+        # Gather response
+        response = requests.post(url, params=payload)
+        if response.status_code != requests.codes.ok:
+            log.error(
+                f"API error {response.status_code}: {response.reason}; {response.content}",
+                function='connexion_lookup',
+            )
+            raise requests.exceptions.HTTPError(response.reason, response.content)
+
+        # Parse response
+        outVal = response.json()
+
+        # Since the gisweb endpoint seems to always return an HTTP 200, also check
+        # the JSON for an 'error' key
+        if 'error' in outVal:
+            errDict = outVal['error']
+            log.error(
+                f"API error {errDict['code']}: {errDict['message']}",
+                function='connexion_lookup',
+            )
+            raise requests.exceptions.HTTPError(errDict['code'], errDict['message'])
+
+        statusInput = outVal['features'][0]['attributes']['INVENTORY_STATUS_CODE']
+
+    except requests.exceptions.HTTPError:
+        return None
+
+    except (IndexError, KeyError):
+        return None
+
+    else:
+        statusInput = statusInput.lower()
+
+        # If we made it to this point, Connexion will be or is currently
+        # available
+        if statusInput in (
+                'released',
+                'out of warranty',
+        ):      # this is the 'available' case
+            return True
+
+        else:
+            return False
+
+
+def gma_lookup(coord_string):
+    """
+    Look up the GMA location given the coordinate string.
+
+    Parameters
+    ----------
+    coord_string : str
+        Formatted <x>,<y> string of coordinates from address_lookup().
+
+    Raises
+    ------
+    requests.exceptions.HTTPError
+        An issue with the lookup endpoint.    
+
+    Returns
+    -------
+    Boolean 'status', designating True for an address within the GMA, or False
+    otherwise.
+
+    """
+
+    url = 'https://gisweb.fcgov.com/arcgis/rest/services/FCMaps/MapServer/26/query'
+
+    payload = {
+        # Manually stringify 'geometry' - requests and json.dumps do this
+        # incorrectly
+        'geometry': """{"points":[["""+coord_string+"""]],"spatialReference":{"wkid":102653}}""",
+        'geometryType': 'esriGeometryMultipoint',
+        'inSR': 2231,
+        'spatialRel': 'esriSpatialRelIntersects',
+        'where': '',
+        'returnGeometry': 'false',
+        'outSR': 2231,
+        'outFields': '*',
+        'f': 'pjson',
+    }
+
+    try:
+        # Gather response
+        response = requests.get(url, params=payload)
+        if response.status_code != requests.codes.ok:
+            log.error(
+                f"API error {response.status_code}: {response.reason}; {response.content}",
+                function='gma_lookup',
+            )
+            raise requests.exceptions.HTTPError(response.reason, response.content)
+
+        # Parse response
+        outVal = response.json()
+
+        # Since the gisweb endpoint seems to always return an HTTP 200, also check
+        # the JSON for an 'error' key
+        if 'error' in outVal:
+            errDict = outVal['error']
+            log.error(
+                f"API error {errDict['code']}: {errDict['message']}",
+                function='gma_lookup',
+            )
+            raise requests.exceptions.HTTPError(errDict['code'], errDict['message'])
+
+        if len(outVal['features']) > 0:
+            return True
+        else:
+            return False
+
+    except requests.exceptions.HTTPError:
+        return False
+
+
+def validate_usps(inobj):
+    if isinstance(inobj, http.request.QueryDict):
+        # Combine fields into Address
+        address = Address(
+            name=" ",
+            address_1=inobj['address'],
+            address_2=inobj['address2'],
+            city=inobj['city'],
+            state=inobj['state'],
+            zipcode=inobj['zipcode'],
+        )
+
+    elif isinstance(inobj, dict):
+        address = Address(**inobj)
+
+    else:
+        raise AttributeError('Unknown validation input')
+
+    usps = USPSApi(settings.USPS_SID, test=True)
+    validation = usps.validate_address(address)
+    outDict = validation.result
+    try:
+        log.info(
+            f"Address dict found: {outDict}",
+            function='validate_usps',
+        )
+        return outDict
+
+    except KeyError:
+        log.exception(
+            "Address could not be found - no guesses",
+            function='validate_usps',
+        )
+        raise
 
 
 def broadcast_email(email):
@@ -553,6 +906,167 @@ def check_if_user_needs_to_renew(user_id):
     return needs_renewal
 
 
+def finalize_application(user, renewal_mode=False, update_user=True):
+    """
+    Finalize the user's application. This is run after all Eligibility Program
+    files are uploaded.
+
+    The defaults are:
+        - not in renewal_mode
+        - do update the user object (last_completed_at, etc)
+    
+    """
+
+    log.debug(
+        f"Entering function with renewal_mode=={renewal_mode}",
+        function='finalize_application',
+        user_id=user.id,
+    )
+
+    # Get all of the user's eligiblity programs and find the one with the lowest
+    # 'ami_threshold' value which can be found in the related
+    # eligiblityprogramrd table
+    lowest_ami = EligibilityProgram.objects.filter(
+        Q(user_id=user.id)
+    ).select_related(
+        'program'
+    ).values(
+        'program__ami_threshold'
+    ).order_by(
+        'program__ami_threshold'
+    ).first()
+
+    # Now save the value of the ami_threshold to the user's household
+    household = Household.objects.get(
+        Q(user_id=user.id)
+    )
+    household.income_as_fraction_of_ami = lowest_ami['program__ami_threshold']
+
+    if renewal_mode:
+        household.is_income_verified = False
+        household.save()
+
+        # Set the user's last_completed_date to now, as well as set the user's
+        # last_renewal_action to null
+        if update_user:
+            user = User.objects.get(id=user.id)
+            user.renewal_mode = True
+            user.last_completed_at = pendulum.now()
+            user.last_renewal_action = None
+            user.save()
+
+        # Get every IQ program for the user that have a renewal_interval_year
+        # in the IQProgramRD table that is not null
+        users_current_iq_programs = IQProgram.objects.filter(
+            Q(user_id=user.id)
+        ).select_related(
+            'program'
+        ).order_by(
+            'program__renewal_interval_year'
+        ).exclude(
+            program__renewal_interval_year__isnull=True
+        )
+
+        # For each element in users_current_iq_programs, delete the program.
+        # Note that each element has a non-null renewal interval
+        for program in users_current_iq_programs:
+            program.renewal_mode = True
+            program.delete()
+
+        # Get the user's eligibility address
+        eligibility_address = AddressRD.objects.filter(
+            id=user.address.eligibility_address_id).first()
+
+        # Now get all of the IQ Programs for which the user is eligible
+        users_iq_programs = get_users_iq_programs(
+            user.id,
+            lowest_ami['program__ami_threshold'],
+            eligibility_address,
+        )
+
+        # For each IQ program the user was previously enrolled, sort
+        # eligibility status into renewal_eligible and renewal_ineligible
+        renewal_eligible = []
+        renewal_ineligible = []
+
+        for program in users_current_iq_programs:
+            if program.program.id in [x.id for x in users_iq_programs]:
+                renewal_eligible.append(program.program.friendly_name)
+            else:
+                renewal_ineligible.append(program.program.friendly_name)
+
+        # For every eligible IQ program, check if the user should be
+        # automatically applied
+        for program in users_iq_programs:
+            # First, check if `program`` is in users_current_iq_programs; if so,
+            # the user is both eligible and were previously applied/enrolled
+            if program.id in [program.program.id for program in users_current_iq_programs]:
+                # Re-apply by creating the element
+                IQProgram.objects.create(
+                    user_id=user.id,
+                    program_id=program.id,
+                )
+
+            # Otherwise, apply if the program has enable_autoapply set to True
+            elif program.enable_autoapply:
+                # Check if the user already has the program in the IQProgram table
+                if not IQProgram.objects.filter(
+                    Q(user_id=user.id) & Q(
+                        program_id=program.id)
+                ).exists():
+                    # If the user doesn't have the program in the IQProgram
+                    # table, then create it
+                    IQProgram.objects.create(
+                        user_id=user.id,
+                        program_id=program.id,
+                    )
+
+        # Return the target page and a dictionary of <session var>: <value>
+        return (
+            'app:dashboard',
+            {
+                'app_renewed': True,
+                'renewal_eligible': sorted(renewal_eligible),
+                'renewal_ineligible': sorted(renewal_ineligible),
+            }
+        )
+    
+    else:
+        household.save()
+
+        if update_user:
+            user.last_completed_at = pendulum.now()
+            user.save()
+
+        # Get the user's eligibility address
+        eligibility_address = AddressRD.objects.filter(
+            id=user.address.eligibility_address_id).first()
+
+        # Now get all of the IQ Programs for which the user is eligible
+        users_iq_programs = get_users_iq_programs(
+            user.id,
+            lowest_ami['program__ami_threshold'],
+            eligibility_address,
+        )
+        # For every IQ program, check if the user should be automatically
+        # enrolled in it if the program has enable_autoapply set to True
+        for program in users_iq_programs:
+            # program is an IQProgramRD object only if the user has not applied
+            if isinstance(program, IQProgramRD) and program.enable_autoapply:
+                IQProgram.objects.create(
+                    user_id=user.id,
+                    program_id=program.id,
+                )
+                log.debug(
+                    f"User auto-applied for '{program.program_name}' IQ program",
+                    function='finalize_application',
+                    user_id=user.id,
+                )
+
+        # Return the target page and an (empty) dictionary of <session var>: <value>
+        return ('app:broadcast', {})
+
+
 def enable_renew_now(user_id):
     """
     Enable the 'Renew Now' button on the dashboard pages
@@ -624,3 +1138,71 @@ def file_validation(
             ', '.join([x.upper() for x in supported_content_types.keys()])
         ),
     )
+
+
+def finalize_address(instance, is_in_gma, has_connexion):
+    """ Finalize the address, given inputs calculated earlier. """
+
+    # Record the service area and Connexion status
+    instance.is_in_gma = is_in_gma
+    instance.is_city_covered = is_in_gma
+    instance.has_connexion = has_connexion
+
+    # Final step: mark the address record as 'verified' and save
+    instance.is_verified = True
+    instance.save()
+
+
+def remove_ineligible_programs(user_id):
+    """
+    Remove programs that a user no longer qualifies for, based on application
+    updates made after the user selected their programs.
+
+    A use-case for this is when a user's documentation is changed from a 30% AMI
+    eligibility program to a 60% AMI program via the admin panel.
+
+    """
+
+    # Get the user object (with updates that were presumably made just prior)
+    user = User.objects.get(id=user_id)
+
+    # Get the user's eligibility address
+    eligibility_address = AddressRD.objects.filter(
+        id=user.address.eligibility_address_id
+    ).first()
+
+    # Get the user's current programs
+    users_iq_programs = IQProgram.objects.filter(user_id=user.id)
+
+    # Get the programs the user is eligible for
+    eligible_programs = get_eligible_iq_programs(
+        user,
+        eligibility_address,
+    )
+
+    # Compare current programs with eligible programs
+    ineligible_current_programs = [
+        x for x in users_iq_programs if x.program not in eligible_programs
+    ]
+
+    # Remove the user from the ineligible program(s), but ONLY IF they're not
+    # currently enrolled
+
+    # If a user is enrolled in any program, raise an exception
+    enrolled_programs = [
+        x for x in ineligible_current_programs if x.is_enrolled
+    ]
+    if len(enrolled_programs) > 0:
+        raise AttributeError(
+            "User is enrolled in {} but would no longer be eligible with the proposed change".format(
+                ', '.join([x.program.program_name for x in enrolled_programs]),
+            )
+        )
+
+    # Delete user from ineligible programs; return message describing the changes
+    msg = []
+    for program in ineligible_current_programs:
+        program.delete()
+        msg.append(f"User was removed from {program.program.program_name}")
+
+    return '; '.join(msg)
